requirements.txt
flask
pika
mysql-connector-python